+++ core/src/main/java/com/hujiugame/qingfeng/animation/Animation.java
@@ -1,0 +1,6 @@
+package com.hujiugame.qingfeng.animation;
+
+public class Animation
+{
+
+}