+++ core/src/main/java/com/hujiugame/qingfeng/animation/task/AnimationTask.java
@@ -1,0 +1,5 @@
+package com.hujiugame.qingfeng.animation.task;
+
+public class AnimationTask
+{
+}